Remove dependency upon curl
Use the url library of Emacs in order to retrieve the redmine
XML data.

This makes the code platform independent as curl is not readily
available on any OS.

Furthermore, the url-insert-file-contents function determines the
encoding of the data and decodes the data in the buffer so that foreign
characters are displayed correctly.